Rename userControllerNew to userRegisterActionsController as it has to co-exist with userController
