+++ server/test/controllers/userControllerTest.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai'
-import { createUser, loginUser, confirmUser, forgotPassword, resetPassword, resendConfirmUser } from '../../src/controllers/userControllerNew';
+import { createUser, loginUser, confirmUser, forgotPassword, resetPassword, resendConfirmUser } from '../../src/controllers/userRegisterActionsController';
 import { assert, spy, useFakeTimers } from 'sinon';
 const sandbox = require('sinon').sandbox.create();
 const User = require('../../src/models/user.js');
